Trypo-Sense: Detecting Punctures in Space Equipment
Trypo comes from the word trypophobia, fear of holes. The function of
the sensor is to detect holes.
We try to emulate the human body ability to detect injuries.
For example a human body gets poked with a pin, he/she will react with a
OUCH.
Thus saves time for the astronaut from manually looking for the air
leaks.